Various minor improvements to the DcmTag class.
Added comparison operators "==" (equal) and "!=" (not equal) to the DcmTag
class, which - in contrast to the inherited comparison operators from the
base class DcmTagKey - also check the private creator identifier (if
present). Introduced two new regression tests that make sure that these
operators work as expected.

Introduced new method getTagKey() to the DcmTag class, which returns a
const reference to the instance, i.e. not a copy as getXTag() does. Also
tried to avoid calls to getXTag() where possible or useful. In the future,
the new method getTagKey() should be used if really needed.