Added support for the new RFC 8996 TLS profiles.
Added support for the new TLS secure transport connection profiles
defined in Supplement 230, The "BCP 195 RFC 8996" TLS profile
and the "Modified BCP 195 RFC 8996" TLS profile. Also implemented
recommendations of RFC 9325, which is now part of BCP 195.
DCMTK now requires OpenSSL 1.0.2 or newer, since the requirements
of RFC 9325 cannot be implemented with older versions.
Note that support for the Modified BCP 195 RFC 8996 TLS profile
will not be available yet, because of limitations in OpenSSL.
The profile requires support for the TLS 1.2 Camellia ciphersuites
in GCM mode, which OpenSSL does not implement. A configure test will
automatically enable support for this profile when DCMTK is compiled
with a future OpenSSL library that implements this.

This closes DCMTK issue #1030.