Updated Autoconf files for recent dcmtls changes.
Added new configure tests needed for the support of Supplement 230,
which has been implemented for the dcmtls module only recently.

When generating the Autoconf files with "config/autoall", it was
noticed that the configure option --runstatedir disappeared. This
seems to be a known issue of Autoconf 2.69 (built from source code).
It will hopefully be fixed when switching to a newer version.