Fixed various bugs in dcmtls and updated docs.
Fixed various bugs in the revised dcmtls module introduced with commit
d269161f7. Completed dcmtls documentation.

The dcmtls module now offers support for the TLS "server name indication"
(SNI) extension, as recommended by RFC 9325. SNI allows a client to
transmit the name of the intended server during TLS negotiation and
allows a virtual server to redirect the request accordingly.
When SNI is active in a DICOM association acceptor, it will reject
incoming network connections when the client has included the SNI
extension in the TLS ClientHello message with an incorrect server name.
Association negotiation will continue when either the correct SNI
name is present, or the SNI extension is absent. Older servers will
simply ignore the SNI extension.

Thanks to Joerg Riesmeier for the testing and bug reports.