[dcmnet] Enable cancelation of connect attempts in requestAssociationTCP
This is implemented using a cancellation callback (not a cancellation flag).
The reason for using a cancellation callback function is that this relieves
the dul implementation from having to deal with potential data races related
to a cancellation flag. It is now up to the caller to ensure appropriate
synchronization of any data being involved in cancellation.