Fixed infinite loop in sendDcmDataset().
Fixed infinite loop in sendDcmDataset() that can occur in rare cases
when sending in Deflated Explicit VR transfer syntax by adding
a dummy call to DcmOutputStream::write() whenever a certain number of
bytes is needed to be available in the output buffer. This causes
the compression process to be run in order to make room in the buffer.

Thanks to Peter Klotz <peter.klotz@siemens-healthineers.com> for the
bug report and sample file.

This closes DCMTK issue #1074.