Updated "dcmrt" classes based on DICOM 2023b.
Updated automatically generated IOD and sequence C++ classes for the
various RT objects based on the current edition of the DICOM standard
(2023b). This is a major update, so there can still be issues.

Also fixed an "ugly" bug when iterating over sequence items using the
gotoNextItem() method of the automatically generated sequence classes.
Thanks to DCMTK forum user "nbeck" for the report.