Fixed error code for unknown TLS ciphersuite.
When an unknown TLS ciphersuite name was passed to one of the command
line tools using the --cipher option, or passed to
DcmTLSCiphersuiteHandler::addCipherSuite() on API level,
an generic "Illegal Call" error was returned. This has now been
replaced with a more descriptive "Unknown or invalid ciphersuite"
error that also reports the name of the unknown ciphersuite.