Better randomness for temporary filenames.
Make seed static so that multiple/concurrent calls to this method
will not use a seed that is initialized by the same time. Instead,
rely on a seed that is updated by each call to makeFilename below, thus
resulting in more "randomness" when called within the same second.

Thanks to GitHub user "bettar" (Alex Bettarini) for the report and patch.