Restored behaviour of INSTALL target on Windows.
By default, the INSTALL target on Windows will now behave as in previous
releases. A new CMake cache variable DCMTK_USE_WIN32_PROGRAMDATA has been
added that allows to enable the experimental new behaviour under which
data files, documentation and configuration files will be installed in
a subdirectory of the Windows ProgramData directory (on most Windows
installations, this is located in C:\ProgramData).