Fixed compile issues on Debian GNU Hurd.
Fixed compile issues on Debian GNU Hurd. In particular, replaced the use
of the PATH_MAX constant with an internal constant named OFICONV_PATH_MAX,
because PATH_MAX is not available on all systems and known to be "broken",
since on many operating systems the maximum path length is determined by
the file system, and not fixed for the entire operating system. We now use
a maximum path length of 4 kBytes, which is the longest PATH_MAX on the
supported operating systems. Note that longer paths will not cause a
buffer overrun, but may cause data files not to be found if the path is
too long.

Thanks to Mathieu Malaterre <mathieu.malaterre@gmail.com> for reporting
this issue.