New implementation of OFStandard::atof().
Wrote a new implementation of OFStandard::atof() that internally uses
C++ iostreams for the conversion of text to double and should thus be
standard conformant on all platforms. The previous implementation, which
was in use since DCMTK 3.5.2, was found to cause rounding errors in
certain use cases. The old implementation can be re-enabled by defining
the macro ENABLE_OLD_OFSTD_ATOF_IMPLEMENTATION.

Thanks to Mathieu Malaterre <mathieu.malaterre@gmail.com> for the bug report.

This closes DCMTK issue #1100.