Build Windows binary for every pushed commit.
This uses GitHub actions to build new DCMTK binaries for Windows
whenever a new commit is pushed to DCMTK/dcmtk master branch.

If the commit is associated with a tag, a new release is produced.
Otherwise, the tag "latest" is created (or updated) with new source
code and binaries.

The new package is only produced when DCMTK successfully builds and
all tests on the build machine succeed as well.

The machine currently used for building is windows/latest with
Visual Studio 17 2022 64 bit.

The build makes use of all DCMTK support libraries, as they are
available at https://dicom.offis.de/download/dcmtk/dcmtk368/support/ .