Extended UID mapping table with properties.
Extended UID mapping table with properties that specify e.g. the standard
that defines the UID, the type and subtype of the UID, and in case of
Storage SOP Classes also the type of the underlying IOD. These properties
can be accessed using the new function dcmGetPropertiesOfUID(). All other
dcmFindXXXOfUID() and dcmFindUIDFromXXX() functions remain unchanged.