Made use of new dcmGetPropertiesOfUID() function.
Made use of the new dcmGetPropertiesOfUID() function for creating a
DICOMDIR. This simplifies the code considerably, as the type of the
Storage SOP Class can now be determined from the UID value. The extended
UID mapping table e.g. also contains information on Composite IODs for
which no Directory Record Type is defined, which is now also checked.

This new implementation also supports a few Storage SOP Classes/IODs that
were missing previously.