Fixed issue with implementation of CP-2084.
Fixed issue with the implementation of CP-2084, which apparently did not
work (as expected). The new code lines introduced with commit 24ebf482c
have never been run due to the way the if-conditions were implemented.

Thanks to Markus Sabin <Markus.Sabin@soft-gate.de> for the report and the
sample file with an instance of the Performed Imaging Agent Administration
SR Storage SOP Class that provoked this incorrect behavior.