New implementation of OFStandard::ftoa().
Wrote a new implementation of OFStandard::ftoa() that internally uses
C++ iostreams for the conversion of double to text and should thus be
standard conformant on all platforms. The previous implementation, which
was in use since DCMTK 3.5.2, was found to cause rounding errors in
certain use cases. The old implementation can be re-enabled by defining
the macro ENABLE_OLD_OFSTD_FTOA_IMPLEMENTATION.

This closes DCMTK issue #860.