Disabled oficonv pass-through mode by default.
The oficonv module offers a pass-through mode in which the libary will
simply copy input to output during a character set conversion if
input and output encoding are the same. This is more efficient, but
means that in invalid byte sequences in the source will remain
undetected and copied to the target, whereas the default behaviour
would detect this and stop the conversion. It is also incompatible
with the behaviour of the GNU iconv library. This behaviour is now
disabled by default. It can be re-enabled by compiling with the
macro DCMTK_ENABLE_ICONV_PASSTHROUGH.