If desired, ignore attribute value errors on write.
Ignore attribute errors in IODComponent (IODModule) if desired by the
user. If ignored, the error is just handled as a warning.

While will we want to write only correct objects, we acknowledge that
in the research field there are lots of broken DICOM objects/images
around that must be worked with. Researchers write tools that make use
of this data e.g. by importing Patient and Study data to create their
own derived objects. For those cases it should be possible to ignore
common errors.

This API has also been exposed to IODCommon (which collects the most
common modules used in nearly every IOD for convenience).

Test cases have been added.