Added destructor for csmapper area object.
Implemented a new function _citrus_csmapper_free() that acts as a
destructor for the global csmapper area object, which is allocated
when iconv() is called for the first time, and never freed.
While this is not a problem as such, memory leak detectors will
complain about unfreed memory. Calling this function at application
exit will avoid this.