Radically simplified oflimits.h.
Radically simplified oflimits.h by unconditionally including <limits>,
which is guaranteed to exist in C++98, and reducing OFnumeric_limits
to a typedef. This eliminates the need for the generation of arith.h,
which required a program (arith.cc) to be compiled and run at
configure time, thus causing problems in cross-compilation setups.