Largely revised base class for Transfer Syntaxes.
Distinguish more clearly between encapsulation and compression of pixel
data. This will be needed for the upcoming support of the still rather
new uncompressed encapsulated Transfer Syntax (introduced with CP-2083).
Also enhanced "support" for Transfer Syntaxes that make use of referenced
pixel data, e.g. using a URL.

Improved API of DcmXfer class to better distinguish the encoding and
compression of pixel data. Some old methods have been deprecated and its
use at various places of the DCMTK will disappear with subsequent commits.
Internally, various boolean struct members have been replaced by more
specific enums, which are also easier to extend.

The changes made should be a good basis for future Transfer Syntaxes that
are expected to be added to the DICOM standard and to the DCMTK over time.

Enhanced performance of one often used DcmXfer constructor and assignment
operator by relying on the explicit order of items in an array. It was
always intended to work like this, but not used so far.

Also added missing API documentation and introduced regression tests for
the main methods and constructors of the DcmXfer class.