Revised handling of STL features at configure time.
Revised the handling of DCMTK_ENABLE_STL and the related configure flags such
as DCMTK_ENABLE_STL_LIST, DCMTK_ENABLE_STL_STRING etc. Previously, enabling
STL features caused a test program to be compiled an run, which makes it
difficult to cross-compile DCMTK. Since the STL implementations can be
considered stable now, the STL features of C++98 will simply be enabled
when requested by DCMTK_ENABLE_STL or a related option, and the STL features
of C++11 will be enabled when requested and the C++11 header files related
to these features are present. The test cases have been moved to a set of
unit tests in commit #e68e43878.