Enhanced implementaton of findTagFromName().
Enhanced implementation and documentation of the findTagFromName()
method. Now, also the tag name and the private creator identifier are
copied from the data dictionary if an attribute was found by its name.
Made clear that private Attributes are also supported to some extent.

Enhanced and fixed formatting of documentation of related classes.