Make sure __cplusplus macro is not set too early.
The __cplusplus macro is used in DCMTK code to check whether the compiler
can handle C++11 and above features (if enabled). VS only sets this macro
if explicitly requested. The CMake code commited in 79bb19 evaluates
too early, whether the macro should be checked (before detecting whether
we want to enable C++11 or above at all), leading to the situtation that
the macro is only set when running CMake at least twice.

Now the necessity for the macro is checked at the right position.