Fix C++11 OFVariant build on VS 2017.
This fixes C++11-enabled build of OFVariant on Visual Studio 2017. The
class failed to compile due to the compiler's incomplete or buggy C++11
support. The commited workaround works right now, though OFVariant's
assignment operator may be updated later as well.

Thanks to Jan Schlamelcher for the analysis and suggested patch.