Temptative fix to bit-preserving option ignoring output directory
Simplest attempt at correcting issue #1122.
When movescu is called with +B (--bit-preserving) option, the output file ends up in the current working directory as opposed to the desired output directory even if you add the -od (--output-directory) option in the command line.

Note that a better approach could be to have a singular writing function and adjust the function signatures to pass whether other processing logic should be skipped. However, that could cause a lot of headaches ensuring all logic paths are checked against the current bit-preserving behavior of outputting the data as-is.

I leave the option to the maintainer if to request further work to address this issue more comprehensively or if the change proposed here is good enough for the project.

The changes here appears to work as intended in my test environment.