Avoid dereferencing zeroth element of empty vector.
class DcmTLSCiphersuiteHandler contains two methods that sort a
list of ciphersuites. This caused an access violation of the list
to be sorted was empty (which could never happen with the command
line tools, but was possible in user code if no security profile
was activated). This is now fixed.

Thanks to Piotr Batko <Piotr.Batko@gehealthcare.com> for the bug report.
This closes GitHub pull request #97.