Made sure that ambiguous VRs are mapped correctly.
Made sure that ambiguous VRs, such as "ox" used internally for "OB" or
"OW", are mapped to an appropriate standard VR when reading a dataset
with the global parser option dcmPreferVRFromDataDictionary being enabled.

Thanks to forum user "andreasb" for the original report and GitHub user
"mrbean-bremen" for the pull request and proposed patch.