Fixed compare() method for text-based VRs.
The polymorphic compare() method can be used to compare DICOM elements,
items, sequences, i.e. the full DcmObject hierarchy. There was a bug in
VRs LT, ST, UR and UT that returned "equal" even if two elements are not
not equal (in case both elements have the same number of values,
same string length but not the same value).

Thanks to Marcel Pham <Marcel.Pham@examion.com> for the report and
analysis.