Added encapsulated uncompressed transfer syntax.
Added initial support for the Encapsulated Uncompressed Explicit VR
Little Endian Transfer Syntax that was introduced with CP-2083.

This initial support includes reading and writing of DICOM files and
data sets that are encoded with this new transfer syntax as well as
basic network support (including some of the command line tools).
What is still missing is the transcoding to and from the three other
uncompressed transfer syntaxes that use the native format for encoding
pixel data.

This partly closes DCMTK Conformance #1000.