Added cstdio based atof and ftoa implementations.
Added implementations of OFStandard::atof() and OFStandard::ftoa() that use
cstdio functions (sscanf and snprintf, or the proprietary Win32 versions
thereof that allow a locale object to be passed) instead of iostreams
since on Windows, the iostream based implementation is about 10 times
slower than the cstdio based implementation. On other platforms, the
iostream based implementation, being the cleaner one, is used by default
since the difference in performance is relatively small there.

Furthermore added a new precision value -2 to OFStandard::ftoa()
that, when used with the default "%g" format, causes the shortest string
to be created that can be converted back to a double without loss of precision.
This is now the mode used by print routines such as
DcmFloatingPointDouble::print() and DSRReferencedTimeOffsetList::print().