Fixed OFStandard::ftoa() implementation using MSVCRT.
The old Microsoft Visual C Runtime (MSVCRT) used in VS 2013 and older
prints 3 exponent digits by default, but offers a function to change
the behavior, which is called now.  This issue does not exist anymore
in the Universal C Runtime used by VS 2015 or newer, but is still
present in the MinGW64 toolchain.