Initial commit of the dcm2img command line tool.
Initial commit of dcm2img, a command line tool designed to replace the older
tools dcm2pnm, dcmj2pnm and dcml2pnm. The command line options are
identical to the older tools, so that dcm2img can serve as a drop-in replacement.
The tool now supports JPEG-LS as a new export format for image files to be written.