Implemented support for DICOM CP 2311.
DICOM CP 2311 has made optional the support of the TLS 1.2 CAMELLIA
ciphersuites in GCM mode, which are not supported by OpenSSL, in the
Modified BCP 195 RFC 8996 TLS Profile. Accordingly, this TLS profile is
now enabled in DCMTK and support for Camellia is included if available in
OpenSSL, which is checked by a feature test.