Fixed definition of ssize_t on Win64.
On platforms where ssize_t is not defined, DCMTK so far defined it as long.
This is incorrect on Win64, where size_t is 64 bit, but long is 32-bit.
On Win64, we now define ssize_t as __int64.

Thanks to Helmut Steiner <helmut@shl.at> for the bug report and proposed fix.

This closes DCMTK issue #1131.