Output debug messages to debug logger.
Output debug messages to the debug and not to the info logger. This reduces
the verboseness of the DICOM dataset parser when the "stop parsing at/after
element" option is used.

Also fixed line indentation of the source code (where appropriate).

Thanks to GitHub user "malaterre" for the report and proposed patch.

This closes GitHub pull request #101.