Fixed leakage of network sockets to child processes.
DCMTK functions that call command line tools (such as the --exec-on-reception
option in storescp) 'leaked' the listen socket and the network association
socket to the child process called via execl(). This has now been fixed by
opening the sockets with the SOCK_CLOEXEC flag on platforms where this is
available, and by calling fcntl() with the FD_CLOEXEC flag otherwise.

Thanks to Christian Wetzel <wetzel@phoenix-pacs.de> for the bug report and
sample code.

This closes DCMTK issue #1078.