Changed name of Status Code "0107H" (CP-2384).
Changed name of Status Code "0107H" from "Attribute list error" to
"Attribute list warning" to clarify it is a warning and this code is
associated with the Status Class "Warning". See CP-2384 for details.

The old define STATUS_N_AttributeListError will still be available for
some time, but should no longer be used and should be replaced by the
new definition STATUS_N_AttributeListWarning.