Improved performance of DcmVR::setVR().
When creating and processing very large amounts of DICOM data sets,
the implementation of DcmVR::setVR() could have led to a performance
bottleneck, so it was improved, e.g. by avoiding the call of strncmp()
for comparing the two-letter VR codes.

Thanks to Jesper Alf Dam <Jesper.Dam@mi.medical.canon> for the report,
the performance analysis and the original patch.