Replaced mutex in OFGlobal by read/write lock.
OFGlobal instances now use a read/write lock instead of a mutex in order
to protect access in multi-threaded implementations. This provides a
significant speed-up if many threads read the global objects concurrently.

Thanks to Jesper Alf Dam <Jesper.Dam@mi.medical.canon> for the report,
the performance analysis and the original patch.