Slightly changed implementation of DcmVR::setVR().
Slightly changed implementation of DcmVR::setVR(): now, internal VRs are
not treated differently, but they are either mapped to an unknown VR
with extended length (4 bytes) or with a 2-byte length field, depending
on the characters used for the VR string.

Also enhanced the documentation of this method and made a few local
variables "const".