Added template specializaton for OFGlobal<OFBool>.
Added template specialization for OFGlobal<OFBool> that uses
std::atomic (C++11) instead of a read-write lock and is lock free
on some platforms, thus offering a performance advantage if many
threads concurrently access the global flag (e.g. while generating UIDs).
The specialization is enabled only if DCMTK is compiled with STL, and
C++11 is available.

Thanks to Jesper Alf Dam <Jesper.Dam@mi.medical.canon> for the report,
the performance analysis and the original patch.