Fix binary segmentations with certain dimensions.
Some binary segmentations where the number of total bits per frame is not
divisible by 8 have been reported to break when serialized into a
dataset.

This commit includes a rewrite of the associated frame packing and
unpacking methods, and a set of tests to ensure that these methods are
correct.

This closes DCMTK issue #1132.

Thanks to Melanie Michels<melanie.michels@snkeos.com> for the report.