Added option that allows to relax GSPS validation.
Added an option to the validation code for Grayscale Softcopy Presentation
States that allows to relax the rule that all images referenced from one
presentation state must belong to the same SOP class. The rule can either
be relaxed to permit related "for presentation" and "for processing"
SOP classes, or this specific check can be completely disabled.

Thanks to Oliver Klerx <klerx@phoenix-pacs.de> for the proposal and patch.