Fixed frame size calculation for special cases.
DcmElement::getUncompressedFrameSize() now takes into account special cases
such as compressed images where Bits Allocated has a value that is not a
multiple of 8, but is overwritten during decompression anyway. In the case
of compressed images, the active decoder is now looked up and its behavior is
taken into account in the calculation.

This closes DCMTK issue #1140.