Fixed data corruption in iconv passthrough mode.
Fixed a data corruption that occured if a string longer than 1024 characters
was run through OFiconv() with source and character set being identical
and the pass-through mode being enabled (i.e. DCMTK compiled with the macro
DCMTK_ENABLE_ICONV_PASSTHROUGH). In this situation, the pointer to the next
unprocessed input character was not updated.

Thanks to DCMTK forum user cschreib for the bug report and test program.

This closes DCMTK issue #1143.