Improved performance of DcmList::seek_to().
Significantly improved the performance of the seek_to() method, in
particular when iterating a very long list using the position of the
items, e.g. a DcmPixelSequence with a huge number of instances of
DcmPixelItem such as in Whole Slide Imaging (WSI) objects. Internally,
the seek_to() method is, e.g., used in the insert(), remove() and
getItem() methods of DcmSequenceOfItems and DcmPixelSequence as well
as in the getElement() and remove() method of DcmItem.

Thanks to GitHub user "reunanen" for the report and proposed patch.

This closes GitHub pull request #114.