Improved handling of JPEG image files.
Improved the handling of JPEG image files as input to img2dcm. Several special
cases such as different types of component subsampling, as well as lossy
JPEG images  in RGB or CMYK color (instead of YCbCr) are now correctly
handled, and rejected if necessary.

This closes DCMTK issue #995.

Thanks to Mathieu Malaterre <mathieu.malaterre@gmail.com> for the suggestion.