Stop cmake with error message if C++ standard is 98.
Stop with an error message if somebody tries to compile DCMTK with a compiler
that either does not support C++11, or is configured via CMAKE_CXX_STANDARD
to only use C++98. For now, compilation with C++98 can be enabled be running
cmake -DDCMTK_PERMIT_CXX98=ON, but this will not work anymore in future
DCMTK releases.