Added "datadir" and "docdir" to CMake config.
Added CMake variables CMAKE_INSTALL_DATADIR and CMAKE_INSTALL_DOCDIR
to the DCMTK configuration that is exported to "DCMTKConfig.cmake".
This is particularly important as both paths usually contain the DCMTK
version number, which might not be known to the project that imports
the DCMTK configuration.

Also added quotation marks to all exported installation subdirectories
so that spaces and other "special" characters are supported.