Fixed issue with icon image of IMAGE content item.
Fixed issue with the optional icon image of an IMAGE content item that
was not copied when the setValue() method was called. Depending on how
the underlying DSRImageReferenceValue class was used, this resulted in
the Icon Image Sequence (0088,0200) missing in the created DICOM dataset.

Also added new regression tests that make sure that the icon image is
created correctly and copied when needed.

Thanks to forum user "pintagliata" for the original report.