Made sure float/double values use a decimal point.
Made sure that floating point and double values use a proper decimal
point when output to a C++ stream. This requires to explicitly set the
"C" locale in case the user changed the default, i.e. the global locale.

This partly closes DCMTK Bug #1153.

Thanks to Mathieu Malaterre <mathieu.malaterre@gmail.com> for the report
on a related issue (that is not yet fixed).