Added initial support for new transfer syntax.
Added initial support for the Deflated Image Frame Compression Transfer
Syntax that was introduced with Supplement 244.

The current support includes reading and writing of DICOM files and
data sets that are encoded with this new transfer syntax as well as
basic network support (including some of the command line tools).
What is still missing is the compression and decompression of this
transfer syntax, i.e. the lossless transcoding to/from the various
uncompressed transfer syntaxes.

This closes DCMTK Conformance #1149.