Don't check for libsndfile in default build.
The public DCMTK does not make any use of libsndfile; therefore there
is no need to check for the library at all.

For CMake, the library is not checked unless DCMTK_WITH_SNDFILE is
enabled by the user.

For autoconf, the presence of libsndfile is only checked if
configured with --with-libsndfile. Only if found the script will
print "checking whether to include libsndfile support...yes" and in
all other cases will (default, --without-libsndfile, --with-libsndfile
but library not found) print "no" instead.