Updated data dictionary for DICOM 2025b.
Updated data dictionary for the latest edition of the DICOM standard, which
has been released only recently.

Since one of the DICOM attributes has been "retired", the name of the
associated DcmTagKey constant also changed. This required minor changes to
the source code of the dcmiod, dcmqrdb, dcmrt and dcmwlm module.