Added support for BulkDataURI in JSON export.
Added support for the generation of BulkDataURIs when exporting DICOM files
to JSON in the dcm2json tool. When enabled, bulk data is stored in files
that are expected to be accessible through a WADO-RS service. A user-defined
URI prefix can be specified for this purpose. Since dcm2json is not a complete
DICOMweb implementation, there is no support (yet) for generating MIME multipart
structures with bulk data references formatted as a UUID.

This closes DCMTK Feature #1151.