Fix writing of Tractography Results Series.
Fix writing of Tractography Results Series which has been omitted in
the past.

Also fixed requirement for Referenced Instance Sequence and added
debug messages.

Added test that demonstrates how to create an Tractography Results
object and is helpful to find problems with the code.

Thanks to forum user "hapap" for the report.