Initial release of the json2dcm command line tool.
Initial release of json2dcm, a tool that converts the contents of a
DICOM dataset in JSON encoding to a binary DICOM file or dataset.
The JSON document is expected to conform to the "DICOM JSON Model"
as defined in DICOM Part 18, Section F. In this initial release,
attribute values referenced through BulkDataURIs are not yet supported.