Fixed off-by-one error when processing an image.
Fixed an off-by-one error when processing an unusual encoding of a DICOM
image, e.g. with partial access to the pixel data. The internal handling
of the input pixel data with a certain combination of VR, Bits Allocated
and Bits Stored (see "case 2c") was missing the final pixel value (for
images with an odd number of bytes) in size and could, therefore, result
in an undefined behavior (incl. an out-of-bounds write to a memory block
on the heap).

Thanks to drak <dr4k.sec@gmail.com> for the report and sample file.