Fixed metaheader generation in dcmodify and dcmconv.
The dcmodify and dcmconv tools incorrectly changed the MediaStorageSOPClassUID
and MediaStorageSOPInstanceUID in the DICOM meta-header when processing
DICOMDIR files. This was caused by the EWM_createNewMeta write mode,
which was introduced as the default behavior for DcmFileFormat::saveFile()
in DCMTK 3.6.6. This option deleted and re-created the entire meta-header
before writing the file. When the SOP Class UID and the SOP Instance UID
were not present in the main dataset, a private SOP class with a
newly generated SOP Instance UID were written. This is now fixed:
MediaStorageSOPClassUID and MediaStorageSOPInstanceUID are preserved
when not present in the main dataset.
Furthermore, dcmodify's "--no-meta-uid" command line option, which was
apparently broken since DCMTK 3.6.6, is now also fixed.

This closes DCMTK issue #1148.