Improved CMake code for static builds.
The variables containing the lists of additional libraries that need to be
linked when a certain static library is linked (such as
LIBPNG_EXTRA_LIBS_STATIC) are now also taken into account on Windows.
Furthermore, added some improvements for static builds on Debian Linux.

Thanks to Khang Trần <khangthk@gmail.com> for the pull request.

This closes DCMTK PR #104.