Added support for Patient's Age to "dcmsr"
Added support for the optional attribute Patient's Age from the Patient
Study Module. This attribute is now written to and read from a DICOM
dataset as well as to/from the dcmsr-specific XML document format (see
"dsr2xml.xsd"). There are also new get and set methods. However, the
Patient's Age is neither printed to the console (see "dsrdump") nor
rendered for the (X)HTML output (see "dsr2html").

Also extended existing test case in order to check whether the get and
set methods work as expected.

Thanks to GitHub user "parasite-lost" (Ulrich Dorsch) for the original
patch, which is almost identical to this commit.

This closes GitHub PR #126.