Added check of "writeMode" to write() method.
Added check of "writeMode" parameter to DcmFileFormat::write() method,
so that only a dataset (without meta header) is written if its value
is EWM_dataset, i.e. if requested by the caller. This behavior is now
consistent with the DcmFileFormat::saveFile() method.

Thanks to Mathieu Malaterre <mathieu.malaterre@gmail.com> for the
original report on this issue.