Fixed leak in SCP Pool class, re-use threads.
In the old implementation busy worker threads have been joined but not
deleted or re-used.

The updated implementation re-uses busy workers by putting them back
into the pool of idle threads. All threads (if ever busy or not) are now
joined and deleted when the pool is being destroyed.

The test has been expanded to use a more thread-safe approach on result
variables and to also tests thread-reuse for threads that had been put
back into the idle pool before.

Thanks to Chuang Zhao <zhaocccchuang@163.com> for the report and
analysis.