Added JPEG decoder option for preserving BitsStored.
Added a codec parameter and corresponing command line options in dcmdjpeg
and dcm2img that allow the value of Bits Stored to be preserved even if
the value in the JPEG bitstream is smaller than the value in the DICOM header.
This may help with correctly decoding some defective DICOM images.

This closes DCMTK issue #1054.