Fixed dcmAcceptOddAttributeLength implementation.
The global dcmdata setting 'dcmAcceptOddAttributeLength' determines how
illegal data elements with odd length are handled when reading.
By default, DCMTK accepts odd-length elements and silently pads
them to even length. If this setting is set to false, DCMTK instead
assumes that the attribute value is in fact even length and that
only the value of the length field is wrong, and increases the
length field by 1. This behaviour, which is not the default,
did not work correctly for elements larger than 4 kBytes, which are
skipped when reading from file and only loaded on demand later.
This has now been fixed.