Fixed issue with VR scanner for "DateTime" values.
Fixed an issue with the VR scanner for an incorrect definition of
"DateTime" (DT) values: The UTC suffix "+0000" was not supported,
which was quite surprising since this is the current timezone for
for the UK, for example.

Also added a new regression test to check whether the VR scanner
correctly accepts and rejects certain timezones for DT values.

Thanks to Jez Cooke <jez.cooke@hamamatsu.eu> and Kade Rashid
<kade.rashid@hamamatsu.eu> for the report and suggested fix.