Added callback parameter to "request fork" function.
Added an optional callback function pointer as a parameter to
DUL_requestForkOnTransportConnectionReceipt(). The callback will
only be used on Windows and on that platform will pass a handle
to the child process that has been created in the network module
to serve the incoming network assocation. The handle can be
used to maintain a list and count of active child processes
on Windows, where no SIGCHLD signal is available that automatically
notifies the parent process when a child process has ended.